Simplify CommandSendStorage
